rollback hpos store
